# Extend sort's floating point optimizations to Float16