Previously, we treated Float16 as an unknown type which we use the standard `lt` function to deal with. This PR extends the optimizations applied to `Float32` and `Float64` to also apply to `Float16`. These optimizations are especially nice because with so few bits, `Float16`s are easily sorted by counting sort or radix sort. The net result is a 30x speedup for large vectors and a smaller speedup for smaller vectors. The increase in allocations is due to use of a different algorithm. The fpsort optimizations are phased out [for vectors smaller than 10](https://github.com/JuliaLang/julia/blob/fb760d973630afc7560be4c225a4a2b2ad1b9fa5/base/sort.jl#L1578-L1580) when they stop being helpful.

<details>

<summary> Benchmarks: </summary>

```julia
for i in 1:8
    x = rand(Float16, 10^i)
    @time sort!(x)
end

# before
  0.000001 seconds
  0.000005 seconds
  0.000081 seconds
  0.000883 seconds
  0.008984 seconds
  0.120891 seconds
  0.684031 seconds
  8.030174 seconds

# after
  0.000003 seconds
  0.000006 seconds (3 allocations: 624 bytes)
  0.000013 seconds (3 allocations: 3.156 KiB)
  0.000101 seconds (4 allocations: 20.703 KiB)
  0.000466 seconds (2 allocations: 120.047 KiB)
  0.004767 seconds (2 allocations: 120.047 KiB)
  0.036061 seconds (2 allocations: 120.047 KiB)
  0.297907 seconds (2 allocations: 120.047 KiB)

for i in 1:8
    @btime sort!(x) setup=(x=rand(Float16, 10^$i)) evals=1
end

# before
  171.000 ns (0 allocations: 0 bytes)
  2.724 μs (0 allocations: 0 bytes)
  40.461 μs (0 allocations: 0 bytes)
  538.888 μs (0 allocations: 0 bytes)
  5.834 ms (0 allocations: 0 bytes)
  63.062 ms (0 allocations: 0 bytes)
  878.109 ms (0 allocations: 0 bytes)
  12.620 s (0 allocations: 0 bytes)

# after
  189.000 ns (0 allocations: 0 bytes)
  1.476 μs (3 allocations: 624 bytes)
  6.955 μs (3 allocations: 3.16 KiB)
  66.651 μs (4 allocations: 20.70 KiB)
  333.517 μs (2 allocations: 120.05 KiB)
  2.643 ms (2 allocations: 120.05 KiB)
  29.109 ms (2 allocations: 120.05 KiB)
  293.027 ms (2 allocations: 120.05 KiB)
```

</details>